203: UI fixes
- Changed default traffic light colour
- Demoted a heading to make it smaller
- Repositioned repository list heading next to button
- Changed the avatar from a square to a circle
- Changed colour of "login" heading to dark grey
- Re-aligned "Add" button to the right on the "select repos" list
- Added cancel button to select repos list
- Added "X selected repos" message to repo select list
- Hid cancel button when no repos are shown
