+++ src/common/components/repositories-home/repositories-home.css
@@ -1,5 +1,5 @@
 .button-container {
   display: flex;
-  justify-content: flex-end;
+  justify-content: space-between;
   margin: 20px 0;
 }
