+++ src/common/components/select-repository-list/styles.css
@@ -6,12 +6,14 @@
 
 .footer {
   display: flex;
+  justify-content: space-between;
   margin: 60px 0;
 }
 
-.right {
+.footer-right {
   display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 400px;
+}
+
+.button-wrapper {
+  margin: 0 5px;
 }
