+++ src/common/components/traffic-lights/traffic-light.css
@@ -1,3 +1,5 @@
+$signal-colour-default: $mid-grey;
+
 .trafficlight {
   display: flex;
   width:80%;
@@ -5,7 +7,7 @@
 }
 
 .signal {
-  background: #666;
+  background: $signal-colour-default;
   border-radius: 50%;
   color: #fff;
   display: inline-block;
@@ -25,7 +27,7 @@
   display: block;
   height: 4px;
   width: 450%;
-  background: #666;
+  background: $signal-colour-default;
   position: absolute;
   top:50%;
   left:100%;
